Also set version in full build
